+++ ThemesManager.cpp
@@ -174,7 +174,7 @@ void ThemesManager::onePluginWillBeRemoved(const PluginsAvailable &plugin)
             if(static_cast<int>(index)<currentPluginIndex)
                 currentPluginIndex--;
             pluginList.erase(pluginList.begin()+index);
-            if((unsigned int)currentPluginIndex>=pluginList.size())
+            if(static_cast<unsigned int>(currentPluginIndex)>=pluginList.size())
             {
                 ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Critical,"plugin is out of inder!");
                 currentPluginIndex=-1;
@@ -194,13 +194,13 @@ QIcon ThemesManager::loadIcon(const std::string &fileName)
 {
     if(currentPluginIndex==-1)
         return QIcon();
-    if(pluginList.at(currentPluginIndex).factory==NULL)
+    if(pluginList.at(static_cast<unsigned int>(currentPluginIndex)).factory==NULL)
     {
         ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Critical,"Try get icon when the factory is not loaded");
         return QIcon();
     }
     ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Notice,"Send interface pixmap: "+fileName);
-    return pluginList.at(currentPluginIndex).factory->getIcon(fileName);
+    return pluginList.at(static_cast<unsigned int>(currentPluginIndex)).factory->getIcon(fileName);
 }
 
 void ThemesManager::allPluginIsLoaded()
@@ -217,7 +217,7 @@ void ThemesManager::allPluginIsLoaded()
     {
         if(pluginList.at(index).plugin.name==name)
         {
-            currentPluginIndex=index;
+            currentPluginIndex=static_cast<int>(index);
             emit theThemeIsReloaded();
             return;
         }
@@ -235,18 +235,18 @@ PluginInterface_Themes * ThemesManager::getThemesInstance()
         ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Warning,"Unable to load the interface, copy aborted");
         return NULL;
     }
-    ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Notice,"Send interface: "+pluginList.at(currentPluginIndex).plugin.name);
-    if((unsigned int)currentPluginIndex>=pluginList.size())
+    ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Notice,"Send interface: "+pluginList.at(static_cast<unsigned int>(currentPluginIndex)).plugin.name);
+    if(static_cast<unsigned int>(currentPluginIndex)>=pluginList.size())
     {
         ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Warning,"Unable to load the interface, internal selection bug");
         return NULL;
     }
-    if(pluginList.at(currentPluginIndex).factory==NULL)
+    if(pluginList.at(static_cast<unsigned int>(currentPluginIndex)).factory==NULL)
     {
         ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Warning,"No plugin factory loaded to get an instance");
         return NULL;
     }
-    return pluginList.at(currentPluginIndex).factory->getInstance();
+    return pluginList.at(static_cast<unsigned int>(currentPluginIndex)).factory->getInstance();
 }
 
 #ifdef ULTRACOPIER_DEBUG
@@ -262,11 +262,11 @@ void ThemesManager::newOptionValue(const std::string &group,const std::string &n
     {
         if(!PluginsManager::pluginsManager->allPluginHaveBeenLoaded())
             return;
-        if(currentPluginIndex!=-1 && value!=pluginList.at(currentPluginIndex).plugin.name)
+        if(currentPluginIndex!=-1 && value!=pluginList.at(static_cast<unsigned int>(currentPluginIndex)).plugin.name)
         {
             //int tempCurrentPluginIndex=currentPluginIndex;
             emit theThemeNeedBeUnloaded();
-            ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Notice,"unload the themes: "+pluginList.at(currentPluginIndex).plugin.name+" ("+std::to_string(currentPluginIndex)+")");
+            ULTRACOPIER_DEBUGCONSOLE(Ultracopier::DebugLevel_Notice,"unload the themes: "+pluginList.at(static_cast<unsigned int>(currentPluginIndex)).plugin.name+" ("+std::to_string(currentPluginIndex)+")");
             /* Themes remain loaded for the options
              *if(pluginList.at(tempCurrentPluginIndex).options!=NULL)
             {
